refactor(codemirror): extract inline widget module
